+++ client/app/components/packages/admin.package.component.ts
@@ -1,14 +1,14 @@
 import { Component, ViewChild, ElementRef, AfterViewInit, OnInit, OnDestroy } from '@angular/core';
 import { Observable } from 'rxjs/Observable';
 import { Router, RouterStateSnapshot } from '@angular/router';
-import { PackageServices,AlertService } from '../../services/index';
+import { PackageServices, AlertService } from '../../services/index';
 
 
 
 @Component({
-  moduleId:module.id,
+  moduleId: module.id,
   selector: 'admin-packages',
-  templateUrl:'./admin.package.component.html'
+  templateUrl: './admin.package.component.html'
   //styles: [main]
 })
 export class AdminPackageComponent implements AfterViewInit {
@@ -43,25 +43,29 @@ export class AdminPackageComponent implements AfterViewInit {
       );
   }
 
-  deletePackage(packageId:any){
-    alert(packageId);
-    this.packageService.deletePackageByid(packageId)
-      .subscribe(data => {
-        if(data.status=='success'){
-          this.alertService.success(data.msg,'packageAlert');
-          this.getAllPackage();
-        }else{
-          this.alertService.error(data.msg,'packageAlert');
+  deletePackage(packageId: any) {
+    var con = confirm("Are you sure!, You want to delete this package");
+    if (con) {
+
+
+      this.packageService.deletePackageByid(packageId)
+        .subscribe(data => {
+          if (data.status == 'success') {
+            this.alertService.success(data.msg, 'packageAlert');
+            this.getAllPackage();
+          } else {
+            this.alertService.error(data.msg, 'packageAlert');
+          }
+        },
+        error => {
+          const body = error.json() || '';
+          const err = body.error || JSON.stringify(body);
+          var errr = JSON.parse(err);
+
+          this.alertService.error(errr.msg, 'packageAlert');
         }
-      },
-      error => {
-        const body = error.json() || '';
-        const err = body.error || JSON.stringify(body);
-        var errr = JSON.parse(err);
-        
-        this.alertService.error(errr.msg,'packageAlert');
-      }
-      );
+        );
+    }
   }
   ngAfterViewInit() {
     this.getAllPackage();
